split createReader in dedicated file
